工作台相关接口

+++ sky-pojo/src/main/java/com/hugo/vo/BusinessDataVO.java
@@ -18,12 +18,12 @@ public class BusinessDataVO implements Serializable {
 
     private Double turnover;//营业额
 
-    private Integer validOrderCount;//有效订单数
+    private Long validOrderCount;//有效订单数
 
     private Double orderCompletionRate;//订单完成率
 
     private Double unitPrice;//平均客单价
 
-    private Integer newUsers;//新增用户数
+    private Long newUsers;//新增用户数
 
 }
