+++ sky-pojo/src/main/java/com/hugo/vo/OrderOverViewVO.java
@@ -16,17 +16,17 @@ import java.io.Serializable;
 @AllArgsConstructor
 public class OrderOverViewVO implements Serializable {
     //待接单数量
-    private Integer waitingOrders;
+    private Long waitingOrders;
 
     //待派送数量
-    private Integer deliveredOrders;
+    private Long deliveredOrders;
 
     //已完成数量
-    private Integer completedOrders;
+    private Long completedOrders;
 
     //已取消数量
-    private Integer cancelledOrders;
+    private Long cancelledOrders;
 
     //全部订单
-    private Integer allOrders;
+    private Long allOrders;
 }
